@@ -1,32 +1,40 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
+  <!-- 3.2.0 using Zeebe 8.5.x/ SporingBoot 3.3.x -->
   <groupId>io.camunda.community</groupId>
   <artifactId>zeebe-cherry-runtime</artifactId>
-  <version>3.1.1</version>
+  <version>3.2.0</version>
 
   <properties>
-    <java.version>17</java.version>
+    <java.version>21</java.version>
     <maven.compiler.target>${java.version}</maven.compiler.target>
     <maven.compiler.source>${java.version}</maven.compiler.source>
 
+    <!-- versin in https://central.sonatype.com/artifact/io.camunda.connector/connector-validation/versions -->
     <!--Version 8.3.0-->
-    <zeebe.version>8.3.0</zeebe.version>
-    <zeebe-client.version>8.3.3</zeebe-client.version>
-    <connector-core.version>8.3.1</connector-core.version>
-    <connector-validation.version>8.3.1</connector-validation.version>
+    <version.zeebe>8.5.7</version.zeebe>
+    <!--Version 8.3.3-->
+    <version.zeebe-client>8.5.5</version.zeebe-client>
+    <!--Version 8.3.1-->
+    <version.connector-core>8.5.4</version.connector-core>
+    <version.connector-runtime>8.5.4</version.connector-runtime>
+    <!--Version 8.3.1-->
+    <version.connector-validation>8.5.4</version.connector-validation>
+
+
 
     <!-- Access file -->
-    <filestorage.version>1.2.0</filestorage.version>
+    <version.filestorage>1.2.0</version.filestorage>
 
     <!-- 2.7.4 -->
-    <spring.boot.version>2.7.4</spring.boot.version>
+    <!-- According https://github.com/camunda-community-hub/spring-zeebe/blob/main/pom.xml -->
+    <version.spring.boot>3.3.0</version.spring.boot>
 
-    <junit.jupiter.version>5.9.1</junit.jupiter.version>
-    <opensagres.version>2.0.3</opensagres.version>
-
-    <opensagres.version>2.0.4</opensagres.version>
+    <version.junit.jupiter>5.9.1</version.junit.jupiter>
+    <version.opensagres>2.0.3</version.opensagres>
 
   </properties>
 
@@ -318,8 +280,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.13.0</version>
         <configuration>
-          <source>17</source>
-          <target>17</target>
+          <source>21</source>
+          <target>21</target>
         </configuration>
       </plugin>
 
