@@ -3,7 +3,7 @@ package io.camunda.cherry.embeddedrunner.ping.connector;
 import io.camunda.cherry.definition.AbstractConnectorInput;
 import io.camunda.cherry.definition.RunnerParameter;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 import java.util.Arrays;
 import java.util.List;
 
